requirements.txt
Instrumental_lib == 0.6
PyQt5_sip == 12.9.0
PyVISA == 1.11.3
PyYAML == 5.4.1
ftd2xx == 1.1.2
matplotlib >= 3.4.2
nidaqmx == 0.5.7
numpy <= 1.21
qasync == 0.22.0
scikit_image >= 0.18.3
scipy == 1.7.1
nicelib >= 0.6
